requirements.txt
google-generative-ai
pydantic
fastapi
python-dotenv

